organize syncing related code
